requirements.txt
"""
Requirements pour l'Editorial Agent IA.
"""

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# LLM & AI
openai==1.3.0
groq==0.4.1
httpx==0.25.0

# Data & Processing
feedparser==6.0.10
aiohttp==3.9.0
langdetect==1.0.9

# Storage & Memory
sqlalchemy==2.0.0

# Frontend
streamlit==1.29.0
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7
